refactor(datasource/rpm): extract repomd helpers
