+++ lib/modules/datasource/rpm/common.ts
@@ -1,0 +1,2 @@
+export const datasource = 'rpm';
+export const repomdXmlFileName = 'repomd.xml';
